api/requirements.txt
asgiref==3.6.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Django==4.2
django-cors-headers==3.14.0
django-restframework==0.0.1
djangorestframework==3.14.0
firebase-admin==6.1.0
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.0
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.53.0
httplib2==0.22.0
idna==3.4
msgpack==1.0.5
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
pytz==2023.3
requests==2.28.2
rsa==4.9
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
